+++ app/src/main/java/com/example/parkingsystem/model/Historico.kt
@@ -1,7 +1,7 @@
 package com.example.parkingsystem.model
 
 data class Historico (
-    val idParque: Int,
+    val idParque: Long,
     val nomeParque: String,
     val entrada: String,
     val saida: String
