+++ app/src/main/java/com/example/parkingsystem/model/User.kt
@@ -1,12 +1,11 @@
 package com.example.parkingsystem.model
 
-
 data class Res(
     val responseStatus: Boolean,
     val response: User
 )
 data class User(
-    val id: Int,
+    val id: Long,
     val email: String,
     val firstName: String,
     val lastName: String,
